Cleanup

+++ zone/zonedb.h
@@ -451,7 +451,7 @@ public:
 	bool SaveCharacterDisc(uint32 character_id, uint32 slot_id, uint32 disc_id);
 	bool SaveCharacterLanguage(uint32 character_id, uint32 lang_id, uint32 value);
 	bool SaveCharacterLeadershipAbilities(uint32 character_id, PlayerProfile_Struct* pp);
-	bool SaveCharacterMaterialColor(uint32 character_id, uint32 slot_id, uint32 color);
+	bool SaveCharacterMaterialColor(uint32 character_id, uint8 slot_id, uint32 color);
 	bool SaveCharacterMemorizedSpell(uint32 character_id, uint32 spell_id, uint32 slot_id);
 	bool SaveCharacterPotionBelt(uint32 character_id, uint8 potion_id, uint32 item_id, uint32 icon);
 	bool SaveCharacterSkill(uint32 character_id, uint32 skill_id, uint32 value);
